pyproject.toml
[tool.poetry]
name = "kiln"
version = "0.1.0"
description = ""
authors = ["Taylor Beck <coder@taylor.id>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.0.123"
openai = "^0.27.2"
python-dotenv = "^1.0.0"
click = "^8.1.3"
scrapeghost = "^0.4.2"
pyaudio = "^0.2.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
